% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_function.R
\name{anova_function}
\alias{anova_function}
\title{anova_function
Function to apply an anova across all response variables. Options for running the model on 1 or 2 independent
variables individually, or on the interaction between two independent variables}
\usage{
anova_function(data, colData, response_variable, Var1, Var2, interaction,
  log_transform)
}
\arguments{
\item{data}{A metabolomics count data frame}

\item{colData}{Meta data dataframe for the metabolomics count data frame}

\item{response_variable}{String of the column header for the response variables, usually "Metabolite"}

\item{Var1}{String of the first independent variable you wish to test}

\item{Var2}{String of the second independent variable you wish to test. Optional parameter}

\item{interaction}{Boolean of TRUE or FALSE for whether or not you wish to model an interaction between
independent variables. Optional parameter}

\item{log_transform}{Boolean of TRUE or FALSE for whether or not you wish to log transform your counts}
}
\description{
anova_function
Function to apply an anova across all response variables. Options for running the model on 1 or 2 independent
variables individually, or on the interaction between two independent variables
}
\examples{
anova_function(data = metabolomics_counts, colData = metadata, response_variable = "Metabolite", Var1 = "Treatment", log_transform = TRUE)
}
