% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assign_KEGG.R
\name{Assign_KEGG}
\alias{Assign_KEGG}
\title{Assign_KEGG
Assigns a KEGG value based off of Metabolite name}
\usage{
Assign_KEGG(data, KEGG_key)
}
\arguments{
\item{data}{A metabolomics data frame with a KEGG column}

\item{KEGG_key}{A data frame from the Omu git repository}
}
\description{
Assign_KEGG
Assigns a KEGG value based off of Metabolite name
}
