% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c57_nos2KO_mouse_countDF}
\alias{c57_nos2KO_mouse_countDF}
\title{c57b6J nos2KO metabolomics count matrix}
\format{A data frame with 668 rows and 36 variables:}
\usage{
c57_nos2KO_mouse_countDF
}
\description{
A dataset containing metabolomics counts for an experiment done using
c57b6J wild type and c57b6J nos2 knockout mice
}
\keyword{datasets}
