% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c57_nos2KO_mouse_metadata}
\alias{c57_nos2KO_mouse_metadata}
\title{c57b6J nos2KO meta data}
\format{A data frame with 29 rows and 4 variables:}
\usage{
c57_nos2KO_mouse_metadata
}
\description{
A a meta data file for the c57b6J metabolomics count matrix
}
\keyword{datasets}
