% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seqs.R
\name{get_seqs}
\alias{get_seqs}
\title{Get nucleotide and amino acid sequences for genes}
\usage{
get_seqs(gene_data)
}
\arguments{
\item{gene_data}{A dataframe with genes from KEGG_gather, with class seqs}
}
\description{
Function that gets nt and aa seqs for gene data from KEGG_gather
}
\examples{

gene_data <- c57_nos2KO_mouse_countDF[(1:2),]

gene_data <- KEGG_gather(gene_data = gene_data)

gene_data <- KEGG_gather(gene_data)

gene_data <- get_seqs(gene_data)
}
