% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_Gather.R
\name{KEGG_gather}
\alias{KEGG_gather}
\title{KEGG_gather()
Method for gathering metadata from KEGG API}
\usage{
KEGG_gather(countDF, sig_threshold)
}
\arguments{
\item{countDF}{A metabolmics count dataframe with a KEGG identifier columns}

\item{sig_threshold}{Numeric. A significance threshold to prune out metabolites}
}
\description{
KEGG_gather()
Method for gathering metadata from KEGG API
}
