% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metabo_Hierarchy.R
\name{Metabo_Hierarchy}
\alias{Metabo_Hierarchy}
\title{Metabo_Hierarchy
Annotates a metabolomics data frame with metabolite hiearchy data}
\usage{
Metabo_Hierarchy(data, Metabolite_Hierarchy, keep_unknowns)
}
\arguments{
\item{data}{A metabolomics data frame with a KEGG column}

\item{keep_unknowns}{FALSE removes unknown metabolites, use TRUE to keep}

\item{KEGG_key}{A data frame from the Omu git page with metabolite hiearchy data in the form of Class, Subclass_1, Subclass_2, Subclass_3, and Subclass_4 columns}
}
\description{
Metabo_Hierarchy
Annotates a metabolomics data frame with metabolite hiearchy data
}
