% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metabo_Path.R
\name{Metabo_Path}
\alias{Metabo_Path}
\title{Metabo_Path
Annotates a metabolomics data frame with metabolite hiearchy data}
\usage{
Metabo_Path(data, KEGG_key)
}
\arguments{
\item{data}{A metabolomics data frame with a KEGG column}

\item{KEGG_key}{A data frame from the Omu git page with metabolite hiearchy data in the form of Class, Subclass_1, Subclass_2, Subclass_3, and Subclass_4 columns}
}
\description{
Metabo_Path
Annotates a metabolomics data frame with metabolite hiearchy data
}
