% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omu_summary.R
\name{omu_summary}
\alias{omu_summary}
\title{omu_summary
Performs t test, standard deviation, standard error, FDR correction, fold change, log2FoldChange. The order effects the fold change values}
\usage{
omu_summary(data, colData, numerator, denominator, response_variable, Factor,
  log_transform, p_adjust)
}
\arguments{
\item{data}{should be a metabolomics count data frame}

\item{colData}{is meta data}

\item{numerator}{is the variable you wish to compare against the denominator, in quotes}

\item{denominator}{see above, in quotes}

\item{response_variable}{the name of the column with your response variables}

\item{Factor}{the column name for your independent variables}

\item{log_transform}{TRUE or FALSE value for whether or not log transformation of data is performed before the t test}

\item{p_adjust}{Method for adjusting the p value, i.e. "BH"}
}
\description{
omu_summary
Performs t test, standard deviation, standard error, FDR correction, fold change, log2FoldChange. The order effects the fold change values
}
\examples{
omu_summary(data = c57_nos2KO_mouse_countDF, colData = c57_nos2KO_mouse_metadata,
numerator = "Strep", denominator = "Mock", response_variable = "Metabolite", Factor = "Treatment",
log_transform = TRUE, p_adjust = "BH")
}
