% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omu_to_DESeq2.R
\name{Omu_to_DESeq2}
\alias{Omu_to_DESeq2}
\title{Omu_to_DESeq2}
\usage{
Omu_to_DESeq2(countMatrix, colData, design, comparison, test, fitType,
  reduced_design, group)
}
\arguments{
\item{countMatrix}{A metabolomics integer class matrix that has undergone feature scaling, as well as hierarchy annotation through the Omu python module.}

\item{colData}{A data frame of character vectors assigning metadata to the discrete variables of the countMatrix. Therefore, ncol of colData must equal nrow of countMatrix.}

\item{design}{The statistical model. I.e. ~ Treatment, ~Treatment + Condition, Treatment + Condition:Treatment (for interaction term), etc. See the DESeq2 manual for an in depth explanation of modeling.}

\item{comparison}{A character list for the comparisons you wish to view, listed as c("Factor of Interest", "Numerator level", "Denominator Level")}

\item{test}{Either "LRT" (likelihood ratio test) or "Wald". For LRT, a reduced design is required along with the full design}

\item{fitType}{"parametric", "local", or "mean"}

\item{reduced_design}{A reduced model missing a factor of interest. For LRT}

\item{group}{TRUE or FALSE value. Use group = TRUE if you want to combine Factors of interest into a single factor consisting of all level combinations. If group = TRUE, design must = ~ group. See Vignette for a more detailed explanation on modeling.}
}
\description{
Takes an Omu count matrix and generates a DESeq2 statistical model
}
