% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_Plot.R
\name{PCA_plot}
\alias{PCA_plot}
\title{Principle component analysis plot
internally performs and ordination and outputs a PCA plot}
\usage{
PCA_plot(data, colData, variable, color, response_variable)
}
\arguments{
\item{data}{Metabolomics count data}

\item{colData}{Metabolomics meta data}

\item{variable}{The independent variable you wish to compare and contrast}

\item{color}{String of what you want to color by. usually should be the same as variable}

\item{response_variable}{String of the response_variable, usually should be "Metabolite"}
}
\description{
Principle component analysis plot
internally performs and ordination and outputs a PCA plot
}
