% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pie_Chart.R
\name{Pie_Chart}
\alias{Pie_Chart}
\title{Pie Chart
Convenience wrapper for the pie chart function in plotly}
\usage{
Pie_Chart(data, type, subplot)
}
\arguments{
\item{data}{Datframe output from Count_Fold_Changes}

\item{type}{string value of either "Increase", "ALL" or "Decrease"}

\item{subplot}{Boolean of TRUE or FALSE. FALSE produces one pie chart of type,
TRUE produces 3 pie charts of "All", "Increase", "Decrease
export
example Pie_Chart(data = your_fold_change_counts, type = "Decrease", subplot = FALSE)}
}
\description{
Pie Chart
Convenience wrapper for the pie chart function in plotly
}
