% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plate_Omelette.R
\name{plate_omelette}
\alias{plate_omelette}
\alias{plate_omelette.rxn}
\alias{plate_omelette.genes}
\title{plate_omelette}
\usage{
plate_omelette(countDF, Matrix)

\method{plate_omelette}{rxn}(countDF)

\method{plate_omelette}{genes}(countDF)
}
\description{
plate_omelette
}
