% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Bar.R
\name{Plot_Bar}
\alias{Plot_Bar}
\title{Plot_Bar
Creates a ggplot2 object using the output file from the Count_Fold_Changes function}
\usage{
Plot_Bar(data, fill, size, color)
}
\arguments{
\item{data}{The output file from Count_Fold_Changes}

\item{fill}{A list of length 2 containing colors for filling the bars, Factors are alphanumeric, so the first color is for the "Decrease" bar while the second is for "Increase"}

\item{size}{A list of 2 numbers for the size of the bar outlines.}

\item{color}{A list of length 2 containing colors for the bar outlines}
}
\description{
Plot_Bar
Creates a ggplot2 object using the output file from the Count_Fold_Changes function
}
\examples{
Plot_Bar(data = data, fill = levels, color = levels, size = levels)
}
\keyword{metabo}
