% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{plot_volcano}
\usage{
plot_volcano(data, column, size, strpattern, fill_list, sig_threshold,
  alpha_list, shape_list, color_list)
}
\arguments{
\item{data}{The output file from the Clean_DESeq_results() function.}

\item{column}{The factor as a list, i.e. column = c("Class")}

\item{size}{Size of points in plot}

\item{strpattern}{A list of levels of the factor you want the plot to focus on, i.e. strpattern = c("Carbohydrates", "Organicacids")}

\item{fill_list}{A list of colors you want your points to be. Levels of a factor are organzed alphabetically. All levels not in the strpattern argument will be set to NA.}

\item{sig_threshold}{An integer. Creates a horizontal dashed line for a significance threshold. i.e. sig_threshold = 0.05. Defaut value is 0.05}

\item{alpha_list}{A list for setting transparency of factor levels. i.e. alpha_list = (1, 0.5, 1)}

\item{shape_list}{A list for setting the shapes for your factor levels. See ggplot2 for an index of shapes}

\item{color_list}{A list of colors for the factor levels. If you choose to use shapes with outlines, this list will set the outline colors.}
}
\description{
This function takes an input data frame from T_Test and creates a volcano plot as a ggplot2 object.
}
\examples{
plot_volcano(data, column = c("Subclass_2"), strpattern = c("SugaralcoholsFig", "SugaracidsFig"),
fill_list = c("black", "hotpink", "cyan"), sig_threshold = 1.301029996, alpha_list = c(0.25, 1, 1), shape_list = c(1, 24, 22),
color_list = c("black", "black", "black"))
plot_volcano(data)
}
