% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_table.R
\name{ra_table}
\alias{ra_table}
\title{ra_table}
\usage{
ra_table(data, variable)
}
\arguments{
\item{data}{data frame object from the count_fold_changes function}

\item{variable}{meta data from count_fold_changes, i.e. "Class"}
}
\description{
Make a relative abundance table from the count_fold_changes function output. For making pie charts
}
\examples{
c57_nos2KO_mouse_countDF <- assign_hierarchy(c57_nos2KO_mouse_countDF, TRUE, "KEGG")

t_test_df <- omu_summary(data = c57_nos2KO_mouse_countDF, colData = c57_nos2KO_mouse_metadata,
numerator = "Strep", denominator = "Mock", response_variable = "Metabolite", Factor = "Treatment",
log_transform = TRUE, p_adjust = "BH")

fold_change_counts <- count_fold_changes(data = t_test_df, "Class",
column = "Class", sig_threshold = 0.05, keep_unknowns = FALSE)

ra_table(data = fold_change_counts, variable = "Class")
}
