% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.metabo.R
\name{read.metabo}
\alias{read.metabo}
\title{read.metabo
wrapper for read.csv that appends the "cpd" class}
\usage{
read.metabo(filepath)
}
\arguments{
\item{filepath}{a file path in "" to your metabolomics count table}
}
\description{
read.metabo
wrapper for read.csv that appends the "cpd" class
}
