% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_Test.R
\name{T_Test}
\alias{T_Test}
\title{T_Test
Performs T_Test, Standard Error, FDR correction, Fold Change, log2FoldChange
The order effects the fold change values}
\usage{
T_Test(data, colData, numerator, denominator, response_variable, Factor,
  log_transform)
}
\arguments{
\item{data}{should be a metabolomics count data frame}

\item{colData}{is meta data}

\item{numerator}{is the variable you wish to compare against the denominator, in quotes}

\item{response_variable}{the name of the column with your response variables}

\item{Factor}{the column name for your independent variables}

\item{denominator:see}{above, in quotes}
}
\description{
T_Test
Performs T_Test, Standard Error, FDR correction, Fold Change, log2FoldChange
The order effects the fold change values
}
