% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_functions.R
\name{transform_metabolites}
\alias{transform_metabolites}
\title{transform_metabolites}
\usage{
transform_metabolites(count_data, func)
}
\arguments{
\item{count_data}{Metabolomics data}

\item{func}{a function to transform metabolites by. can be an anonymous function}
}
\description{
A functional to transform metabolomics data by metabolites
}
\examples{
data_pareto_scaled <- transform_samples(count_data = c57_nos2KO_mouse_countDF,
function(x) x/sqrt(sd(x)))
}
