% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_functions.R
\name{transform_samples}
\alias{transform_samples}
\title{transform_samples}
\usage{
transform_samples(count_data, func)
}
\arguments{
\item{count_data}{Metabolomics data}

\item{func}{a function to transform samples by. can be an anonymous function}
}
\description{
A functional to transform metabolomics data by samples
}
\examples{
data_ln <- transform_samples(count_data = c57_nos2KO_mouse_countDF, log)
}
