% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count_Fold_Changes.R
\name{count_fold_changes}
\alias{count_fold_changes}
\title{count_fold_changes}
\usage{
count_fold_changes(data, ..., column, sig_threshold)
}
\arguments{
\item{data}{Output dataframe from the Clean_DESeq_results function}

\item{...}{Either a Class or Subclass column listed in paretheses, i.e. "Class}

\item{column}{The same value entered for the ... parameter, i.e. column = "Class}

\item{alpha}{Significance threshold, i.e. alpha = 0.05}
}
\description{
This function takes an input data frame that has been run in a statistical modeling function
and returns the number of compounds that significantly changed in each metabolite Class or Subclass.
}
\examples{
count_fold_changes(data = data, "Class", column = "Class", alpha = 0.05)
}
\keyword{metabo}
