% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count_Fold_Changes.R
\name{Count_Fold_Changes}
\alias{Count_Fold_Changes}
\title{Count Fold Changes}
\usage{
Count_Fold_Changes(data, ..., column, alpha)
}
\arguments{
\item{data}{Output dataframe from the Clean_DESeq_results function}

\item{...}{Either a Class or Subclass column listed in paretheses, i.e. "Class}

\item{column}{The same value entered for the ... parameter, i.e. column = "Class}

\item{alpha}{Significance threshold, i.e. alpha = 0.05}
}
\description{
This function takes an input data frame that has been run in a statistical modeling function 
and returns the number of compounds that significantly changed in each metabolite Class or Subclass.
}
\examples{
Count_Fold_Changes(data = data, "Class", column = "Class", alpha = 0.05)
Count_Fold_Changes()
}
\keyword{metabo}
