% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_gather.R
\name{KEGG_gather}
\alias{KEGG_gather}
\alias{KEGG_gather.cpd}
\alias{KEGG_gather.rxn}
\alias{KEGG_gather.KO}
\title{KEGG_gather
Method for gathering metadata from KEGG API}
\usage{
KEGG_gather(count_data)

\method{KEGG_gather}{cpd}(count_data)

\method{KEGG_gather}{rxn}(count_data)

\method{KEGG_gather}{KO}(count_data)
}
\arguments{
\item{count_data}{A metabolmics count dataframe with a KEGG identifier columns}
}
\description{
KEGG_gather
Method for gathering metadata from KEGG API
}
