% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_Functions.R
\name{make_omelette}
\alias{make_omelette}
\title{Get metadata from KEGG API}
\usage{
make_omelette(count_data, column, req)
}
\arguments{
\item{count_data}{The metabolomics count data}

\item{column}{The name of the KEGG identifier being sent to the KEGG API}

\item{req}{The character vector with variables to pull out of the metadata}
}
\description{
Internal function for KEGG_Gather
}
