% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_Functions.R
\name{make_omelette}
\alias{make_omelette}
\title{make_omelette
Internal function for KEGG_Gather}
\usage{
make_omelette(countDF, column, req)
}
\description{
make_omelette
Internal function for KEGG_Gather
}
