% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_chart.R
\name{pie_chart}
\alias{pie_chart}
\title{pie_chart
Makes pie chart as ggplot2 object from ra_table function output}
\usage{
pie_chart(data, variable, column, color)
}
\arguments{
\item{data}{a dataframe object of percents. output from ra_table function}

\item{variable}{The meta data variable you are measuring, i.e. "Class"}

\item{column}{either "Increase", "Decrease", or "Significant_Changes"}

\item{color}{string denoting color for outline. use NA for no outline}
}
\description{
pie_chart
Makes pie chart as ggplot2 object from ra_table function output
}
\examples{
c57_nos2KO_mouse_countDF <- assign_hierarchy(c57_nos2KO_mouse_countDF, TRUE, "KEGG")

t_test_df <- t_test(data = c57_nos2KO_mouse_countDF, colData = c57_nos2KO_mouse_metadata,
numerator = "Strep", denominator = "Mock", response_variable = "Metabolite", Factor = "Treatment",
log_transform = TRUE)

fold_change_counts <- count_fold_changes(data = t_test_df, "Class",
column = "Class", sig_threshold = 0.05)

ra_table(data = fold_change_counts, variable = "Class")

pie_chart(data = ra_table, variable = "Class", column = "Decrease", color = "black")
}
