% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_chart.R
\name{pie_chart}
\alias{pie_chart}
\title{pie_chart
Makes pie chart as ggplot2 object from ra_table function output}
\usage{
pie_chart(data, variable, column, color)
}
\arguments{
\item{data}{a dataframe object of percents. output from ra_table function}

\item{variable}{The meta data variable you are measuring, i.e. "Class"}

\item{column}{either "Increase", "Decrease", or "Significant_Changes"}

\item{string}{denoting color for outline. use NA for no outline}
}
\description{
pie_chart
Makes pie chart as ggplot2 object from ra_table function output
}
