% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_omelette.R
\name{plate_omelette}
\alias{plate_omelette}
\alias{plate_omelette.rxn}
\alias{plate_omelette.genes}
\title{plate_omelette
Internal method for KEGG_Gather}
\usage{
plate_omelette(count_data)

\method{plate_omelette}{rxn}(count_data)

\method{plate_omelette}{genes}(count_data)
}
\arguments{
\item{count_data}{The metabolomics count dataframe}
}
\description{
plate_omelette
Internal method for KEGG_Gather
}
