% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_omelette.R
\name{plate_omelette}
\alias{plate_omelette}
\alias{plate_omelette.rxn}
\alias{plate_omelette.genes}
\alias{plate_omelette.KO}
\title{plate_omelette
Internal method for KEGG_Gather which parses flat text files}
\usage{
plate_omelette(output)

\method{plate_omelette}{rxn}(output)

\method{plate_omelette}{genes}(output)

\method{plate_omelette}{KO}(output)
}
\arguments{
\item{output}{The metabolomics count dataframe}
}
\description{
plate_omelette
Internal method for KEGG_Gather which parses flat text files
}
