% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_Functions.R
\name{plate_omelette_rxnko}
\alias{plate_omelette_rxnko}
\title{Plate_Omelette_rxnKO
Internal function for KEGG_Gather.rxn method
KEGG_Gather.rxn requires dispatch on multiple elements, so
There was no way to incorporate as a method}
\usage{
plate_omelette_rxnko(countDF, matrix)
}
\description{
Plate_Omelette_rxnKO
Internal function for KEGG_Gather.rxn method
KEGG_Gather.rxn requires dispatch on multiple elements, so
There was no way to incorporate as a method
}
