% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_Functions.R
\name{plate_omelette_rxnko}
\alias{plate_omelette_rxnko}
\title{Clean up orthology metadata}
\usage{
plate_omelette_rxnko(count_data, matrix)
}
\arguments{
\item{count_data}{Metabolomics count data}

\item{matrix}{the matrix of KEGG metadata}
}
\description{
Internal function for KEGG_Gather.rxn method
KEGG_Gather.rxn requires dispatch on multiple elements, so
There was no way to incorporate as a method
}
