% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Bar.R
\name{plot_bar}
\alias{plot_bar}
\title{plot_bar
Creates a ggplot2 object using the output file from the count_fold_changes function}
\usage{
plot_bar(data, fill, size, color)
}
\arguments{
\item{data}{The output file from Count_Fold_Changes}

\item{fill}{A character vector of length 2 containing colors for filling the bars, Factors are alphanumeric, so the first color is for the "Decrease" bar while the second is for "Increase"}

\item{size}{A character vector of 2 numbers for the size of the bar outlines.}

\item{color}{A character vector of length 2 containing colors for the bar outlines}
}
\description{
plot_bar
Creates a ggplot2 object using the output file from the count_fold_changes function
}
\examples{
c57_nos2KO_mouse_countDF <- assign_hierarchy(c57_nos2KO_mouse_countDF, TRUE, "KEGG")
t_test_df <- t_test(data = c57_nos2KO_mouse_countDF, colData = c57_nos2KO_mouse_metadata,
numerator = "Strep", denominator = "Mock", response_variable = "Metabolite", Factor = "Treatment",
log_transform = TRUE)
fold_change_counts <- count_fold_changes(data = t_test_df, "Class",
column = "Class", sig_threshold = 0.05)
plot_bar(data = fold_change_counts, fill = c("firebrick2", "dodgerblue2"),
color = c("black", "black"), size = c(1,1))
}
