% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{plot_volcano}
\usage{
plot_volcano(data, column, strpattern, fill, sig_threshold, alpha, shape, color)
}
\arguments{
\item{data}{The output file from the T_Test().}

\item{column}{A character vector, i.e. column = c("Class")}

\item{strpattern}{A character vector of levels of the factor you want the plot to focus on, i.e. strpattern = c("Carbohydrates", "Organicacids")}

\item{fill}{A character vector of colors you want your points to be. Levels of a factor are organzed alphabetically. All levels not in the strpattern argument will be set to NA.}

\item{sig_threshold}{Numeric. Creates a horizontal dashed line for a significance threshold. i.e. sig_threshold = 0.05. Defaut value is 0.05}

\item{alpha}{A character vector for setting transparency of factor levels. i.e. alpha = (1, 0.5, 1)}

\item{shape}{A character vector for setting the shapes for your factor levels. See ggplot2 for an index of shapes}

\item{color}{A character vector of colors for the factor levels. If you choose to use shapes with outlines, this list will set the outline colors.}

\item{factor}{A column of metadata from the data file, i.e. data$Class. If you provide the function with a factor parameter, every other parameter except sig_threshold is required. Choosing no factor parameter will produce a volcano plot with significant points in red, and not significant points in black.}
}
\description{
This function takes a data frame from T_Test and creates a Volcano Plot as a ggplot2 graphical object.
}
