% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_table.R
\name{ra_table}
\alias{ra_table}
\title{ra_table}
\usage{
ra_table(data, variable)
}
\arguments{
\item{data}{data frame object from the count_fold_changes function}

\item{variable}{meta data from count_fold_changes, i.e. "Class"}
}
\description{
Make a relative abundance table from the count_fold_changes function output. For making pie charts
}
